change samples to saved object (#427)
* change samples to saved object


* add sample to so


* fix saved object sample


* fix up


* add back toast


* revert file


* refactoring


* use constant


---------

(cherry picked from commit ce06ccb9fc4acd75d3ccbfbacab9db8a521de252)
